rename the collection package to collect
