+++ support/collect/collect.go
@@ -1,4 +1,4 @@
-package collection
+package collect
 
 import (
 	"github.com/samber/lo"
